# backend/smart-space-booking/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/diskon": {
      "get": {
        "operationId": "AdminDiskonController_daftar",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Daftar seluruh kode promo milik admin",
        "tags": [
          "Admin - Kode Promo"
        ]
      },
      "post": {
        "operationId": "AdminDiskonController_buat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiskonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Tambah kode promo atau event diskon baru",
        "tags": [
          "Admin - Kode Promo"
        ]
      }
    },
    "/api/admin/diskon/{id}": {
      "get": {
        "operationId": "AdminDiskonController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Detail kode promo berdasarkan ID",
        "tags": [
          "Admin - Kode Promo"
        ]
      },
      "put": {
        "operationId": "AdminDiskonController_perbarui",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiskonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Perbarui kode promo dan periode berlakunya",
        "tags": [
          "Admin - Kode Promo"
        ]
      },
      "delete": {
        "operationId": "AdminDiskonController_hapus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Hapus kode promo (soft delete)",
        "tags": [
          "Admin - Kode Promo"
        ]
      }
    },
    "/api/admin/members": {
      "get": {
        "operationId": "AdminMembersController_daftar",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Cari berdasarkan nama, instansi, atau nomor telepon",
            "schema": {
              "example": "Budi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Daftar seluruh member",
        "tags": [
          "Admin - Member"
        ]
      },
      "post": {
        "operationId": "AdminMembersController_buat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Tambah data member baru beserta akun loginnya",
        "tags": [
          "Admin - Member"
        ]
      }
    },
    "/api/admin/members/{id}": {
      "get": {
        "operationId": "AdminMembersController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Detail member berdasarkan ID",
        "tags": [
          "Admin - Member"
        ]
      },
      "put": {
        "operationId": "AdminMembersController_perbarui",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Perbarui data member, termasuk reset password",
        "tags": [
          "Admin - Member"
        ]
      },
      "delete": {
        "operationId": "AdminMembersController_hapus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Hapus data member (soft delete)",
        "tags": [
          "Admin - Member"
        ]
      }
    },
    "/api/admin/reservasi": {
      "get": {
        "operationId": "AdminReservasiController_daftar",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 12,
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "example": 2026,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "belum_dikonfirm",
                "disetujui",
                "aktif",
                "selesai",
                "dibatalkan"
              ]
            }
          },
          {
            "name": "id_space",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "tanggal",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-08-30",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Seluruh reservasi pada lokasi ini, dengan filter",
        "tags": [
          "Admin - Reservasi"
        ]
      }
    },
    "/api/admin/reservasi/{id}/status": {
      "patch": {
        "operationId": "AdminReservasiController_ubahStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservasiStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Konfirmasi atau ubah status pemesanan",
        "tags": [
          "Admin - Reservasi"
        ]
      }
    },
    "/api/admin/reservasi/{id}/check-in": {
      "post": {
        "operationId": "AdminReservasiController_checkIn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Check-in pelanggan",
        "tags": [
          "Admin - Reservasi"
        ]
      }
    },
    "/api/admin/reservasi/{id}/check-out": {
      "post": {
        "operationId": "AdminReservasiController_checkOut",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Check-out pelanggan",
        "tags": [
          "Admin - Reservasi"
        ]
      }
    },
...
    "/api/diskon/{id}": {
      "get": {
        "operationId": "DiskonController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Detail diskon berdasarkan ID",
        "tags": [
          "Katalog Diskon"
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Space Booking API",
    "description": "REST API sistem reservasi coworking space untuk UKK RPL 2026/2027 Paket B. Mencakup autentikasi multi-role, katalog space, kode promo, reservasi, e-ticket ber-QR, check-in/check-out, serta rekapitulasi pendapatan bulanan.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Tempelkan access_token hasil login tanpa awalan \"Bearer\""
      }
    },
    "schemas": {
...
      "CreateDiskonDto": {
        "type": "object",
        "properties": {
          "nama_diskon": {
            "type": "string",
            "example": "PROMOAGUSTUS"
          },
          "persentase_diskon": {
            "type": "number",
            "example": 20,
            "minimum": 1,
            "maximum": 100
          },
          "tanggal_awal": {
            "type": "string",
            "example": "2026-08-01T00:00:00Z"
          },
          "tanggal_akhir": {
            "type": "string",
            "example": "2026-08-31T23:59:59Z"
          }
        },
        "required": [
          "nama_diskon",
          "persentase_diskon",
          "tanggal_awal",
          "tanggal_akhir"
        ]
      },
      "UpdateDiskonDto": {
        "type": "object",
        "properties": {
          "nama_diskon": {
            "type": "string",
            "example": "PROMOAGUSTUS"
          },
          "persentase_diskon": {
            "type": "number",
            "example": 20,
            "minimum": 1,
            "maximum": 100
          },
          "tanggal_awal": {
            "type": "string",
            "example": "2026-08-01T00:00:00Z"
          },
          "tanggal_akhir": {
            "type": "string",
            "example": "2026-08-31T23:59:59Z"
          }
        }
      },
      "CreateMemberAdminDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "user_budi"
          },
          "password": {
            "type": "string",
            "example": "Secret123!",
            "minLength": 6
          },
          "nama_member": {
            "type": "string",
            "example": "Budi Raharjo"
          },
          "instansi": {
            "type": "string",
            "example": "SMK Telkom Malang"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Danau Ranau No. 1, Sawojajar, Malang"
          },
          "telp": {
            "type": "string",
            "example": "085712345678"
          },
          "foto": {
            "type": "string",
            "example": "budi_raharjo.jpg"
          }
        },
        "required": [
          "username",
          "password",
          "nama_member",
          "instansi",
          "alamat",
          "telp"
        ]
      },
      "UpdateMemberAdminDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "Secret123!",
            "minLength": 6
          },
          "nama_member": {
            "type": "string",
            "example": "Budi Raharjo"
          },
          "instansi": {
            "type": "string",
            "example": "SMK Telkom Malang"
          },
          "alamat": {
            "type": "string",
            "example": "Jl. Danau Ranau No. 1, Sawojajar, Malang"
          },
          "telp": {
            "type": "string",
            "example": "085712345678"
          },
          "foto": {
            "type": "string",
            "example": "budi_raharjo.jpg"
          }
        }
      },
      "UpdateReservasiStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "belum_dikonfirm",
              "disetujui",
              "aktif",
              "selesai",
              "dibatalkan"
            ],
            "example": "disetujui"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}